* Support now added for multi-arch CDs/DVDs, including mixed
    binary/source discs. Multi-arch discs should be bootable on those
    multiple arches, modulo boot-sector clashes. Extra support added
    in boot-i386 and boot-amd64 to make the 2 main arches happily
    co-exist.

  * Disc sizing is now much more intelligent - sizes are defined
    depending on the disk type chosen at the start, and discs are now
    filled exactly to those sizes while files are added rather than
    the old up-front guessing method. Equally, the metadata on the
    disc (Packages, Packages.gz, Sources, Sources.gz, md5sums.txt) is
    generated as much as possible while this is happening to make the
    sizing code incredibly accurate. Using this method of disc sizing
    means that customising discs should be much easier/safer - either
    add custom contents at the start and debian-cd will fill the
    remainder of the disc, or afterwards roll back the packages on the
    disc and add extras later.

  * Source is now treated as (almost) just another arch, with most of
    the special casing for source hidden internally. If asking for
    source-only discs, they will simply be built using all the
    available sources. If combined with other arches, the sources will
    automatically be chosen to match the binary packages. Meeting GPL
    requirements was never so easy!

  * Removed lots of old cruft to clean up the codebase:
    + non-US support
    + woody support
    + lots of old support scripts that have been made redundant
    + significantly simpler Makefile, much easier to follow
    + old boot-$ARCH.calc files for estimating boot-file sizes are now
      (obviously) obsolete and therefore gone

+++ tools/etch/installtools.sh
@@ -9,47 +9,65 @@ fi
 
 # The location of the tree for CD#1, passed in
 DIR=$1
+ARCHES="$2"
+NUM_ARCHES=`echo $ARCHES | wc -w`
 
 if [ "$OMIT_MANUAL" != 1 ]; then
 	DOCDIR=doc
-	MANTDIR=$TDIR/installguide
+    MANTDIR=$TDIR/installguide
 
-	INSTALLGUIDE=$(zcat $MIRROR/dists/etch/main/binary-$ARCH//Packages.gz | \
-	    sed -n "s/Filename: \(pool\/main\/i\/installation-guide\/installation-guide-$ARCH.*deb\)$/\1/p")
-
-	if [ -f "$MIRROR/$INSTALLGUIDE" ]; then
-		rm -rf $MANTDIR
-		# Extract documentation from package
-		dpkg -x $MIRROR/$INSTALLGUIDE $MANTDIR || true
-
-		if [ -d $MANTDIR/usr/share/doc/installation-guide-$ARCH/ ]; then
-			cd $MANTDIR/usr/share/doc/installation-guide-$ARCH
-			rm -f changelog* copyright
-			gunzip *.gz || true
-
-			mkdir -p $DIR/$DOCDIR/manual
-			if ! cp -a * $DIR/$DOCDIR/manual; then
-				echo "ERROR: Unable to copy installer documentation to CD."
+	for ARCH in $ARCHES
+	do
+		if [ $ARCH != source ] ; then
+			if [ $NUM_ARCHES = 1 ] ; then
+				INSTALLDIR=$DIR/$DOCDIR/manual
+			else
+				INSTALLDIR=$DIR/$DOCDIR/manual/$ARCH
 			fi
-		else
-			echo "ERROR: installlation-guide package not unpacked correctly."
+
+	        DOCS=$MIRROR/dists/$DI_CODENAME/main/installer-$ARCH/current/$DOCDIR
+            INSTALLGUIDE=$(zcat $MIRROR/dists/etch/main/binary-$ARCH//Packages.gz | \
+                sed -n "s/Filename: \(pool\/main\/i\/installation-guide\/installation-guide-$ARCH.*deb\)$/\1/p")
+
+            if [ -f "$MIRROR/$INSTALLGUIDE" ]; then
+                rm -rf $MANTDIR
+                # Extract documentation from package
+                dpkg -x $MIRROR/$INSTALLGUIDE $MANTDIR || true
+                if [ -d $MANTDIR/usr/share/doc/installation-guide-$ARCH/ ]; then
+                    cd $MANTDIR/usr/share/doc/installation-guide-$ARCH
+                    rm -f changelog* copyright
+                    gunzip *.gz || true
+
+                    mkdir -p $INSTALLDIR
+                    if ! cp -a * $INSTALLDIR; then
+                        echo "ERROR: Unable to copy installer documentation to CD."
+                    fi
+                else
+                    echo "ERROR: installation-guide package not unpacked correctly."
+               fi
+            else
+                echo "ERROR: package installation-guide-$ARCH not found."
+            fi
 		fi
-	else
-		echo "ERROR: package installlation-guide-$ARCH not found."
-	fi
+	done
 fi
 
 if [ "$OMIT_RELEASE_NOTES" != 1 ]; then
-	RN=$DIR/doc/release-notes
-	mkdir -p $RN
-	cd $RN
-	echo "Downloading most recent release notes"
-	wget $RELEASE_NOTES_LOCATION/release-notes-$ARCH.tar.gz
-	if [ -e release-notes-$ARCH.tar.gz ] ; then
-		tar xzvf release-notes-$ARCH.tar.gz
-		rm -f release-notes-$ARCH.tar.gz
-		rm -f */*.ps
-	else
-		echo "No release notes found at $RELEASE_NOTES_LOCATION/release-notes-$ARCH.tar.gz"
-	fi
+	for ARCH in $ARCHES
+	do
+		if [ $ARCH != source ] ; then
+			RN=$DIR/doc/release-notes
+			mkdir -p $RN
+			cd $RN
+			echo "Downloading most recent release notes for $ARCH"
+			wget $RELEASE_NOTES_LOCATION/release-notes-$ARCH.tar.gz
+			if [ -e release-notes-$ARCH.tar.gz ] ; then
+				tar xzvf release-notes-$ARCH.tar.gz
+				rm -f release-notes-$ARCH.tar.gz
+				rm -f */*.ps
+			else
+				echo "No release notes found at $RELEASE_NOTES_LOCATION/release-notes-$ARCH.tar.gz"
+			fi
+		fi
+	done
 fi
